added SharedMenuButton external comp
