+++ libs/shared/menu-button/README.md
@@ -1,0 +1,7 @@
+# shared-menu-button
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test shared-menu-button` to execute the unit tests via [Jest](https://jestjs.io).
