+++ libs/shared/menu-button/jest.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  name: 'shared-menu-button',
+  preset: '../../../jest.config.js',
+  transform: {
+    '^.+\\.[tj]sx?$': 'ts-jest',
+  },
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'html'],
+  coverageDirectory: '../../../coverage/libs/shared/menu-button',
+};
