+++ libs/shared/menu-button/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/shared-menu-button';
