+++ libs/shared/menu-button/src/lib/shared-menu-button.spec.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { render } from '@testing-library/react';
+
+import SharedMenuButton from './shared-menu-button';
+
+describe(' SharedMenuButton', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<SharedMenuButton />);
+    expect(baseElement).toBeTruthy();
+  });
+});
