+++ tsconfig.json
@@ -18,7 +18,8 @@
     "paths": {
       "@ppm/shared/storybook": ["libs/shared/storybook/src/index.ts"],
       "@ppm/shared/test": ["libs/shared/test/src/index.ts"],
-      "@ppm/shared/advert-card": ["libs/shared/advert-card/src/index.ts"]
+      "@ppm/shared/advert-card": ["libs/shared/advert-card/src/index.ts"],
+      "@ppm/shared/menu-button": ["libs/shared/menu-button/src/index.ts"]
     }
   },
   "exclude": ["node_modules", "tmp"]
